update user management with branch

+++ app/Http/Controllers/Web/UserManagement/UserManagementController.php
@@ -4,9 +4,11 @@ namespace App\Http\Controllers\Web\UserManagement;
 
 use App\Http\Controllers\BaseController;
 use App\Http\Resources\PaginationResource;
+use App\Jobs\SendEmailJob;
 use App\Mail\UserManagerVerification;
 use App\Models\User;
 use App\Models\UserManager;
+use App\Repositories\RolePermissionManager\RolePermissionManagerInterface;
 use App\Repositories\UserManagement\UserManagementInterface;
 use App\Repositories\UserVerification\UserVerificationInterface;
 use Illuminate\Http\Request;
@@ -17,56 +19,66 @@ use Illuminate\Support\Facades\Validator;
 
 class UserManagementController extends BaseController
 {
-    public $userManagement, $userVerification;
+    public $userManagement, $userVerification,$roles;
 
     /**
      * @param App\Repositories\UserManagement\UserManagementInterface $userManagement
      * @param App\Repositories\UserVerification\UserVerificationInterface $userVerification
      * @return null
      */
-    public function __construct(UserManagementInterface $userManagement, UserVerificationInterface $userVerification)
-    {
+    public function __construct(
+        UserManagementInterface $userManagement,
+        UserVerificationInterface $userVerification,
+        RolePermissionManagerInterface $roles
+    ) {
         $this->userManagement = $userManagement;
         $this->userVerification = $userVerification;
-        $this->middleware('userpermissionmanager:user-management-permission-list',['only' => 'index']);
-        $this->middleware('userpermissionmanager:user-management-permission-detail',['only' => 'detail']);
-        $this->middleware('userpermissionmanager:user-management-permission-create',['only' => 'create']);
-        $this->middleware('userpermissionmanager:user-management-permission-update',['only' => 'update']);
-        $this->middleware('userpermissionmanager:user-management-permission-delete',['only' => 'delete']);
-        $this->middleware('userpermissionmanager:user-management-permission-resend',['only' => 'resendInvitation']);
+        $this->roles = $roles;
+        // $this->middleware('userpermissionmanager:user-management-permission-list', ['only' => 'index']);
+        // $this->middleware('userpermissionmanager:user-management-permission-detail', ['only' => 'detail']);
+        // // $this->middleware('userpermissionmanager:user-management-permission-create', ['only' => 'create']);
+        // $this->middleware('userpermissionmanager:user-management-permission-update', ['only' => 'update']);
+        // $this->middleware('userpermissionmanager:user-management-permission-delete', ['only' => 'delete']);
+        // $this->middleware('userpermissionmanager:user-management-permission-resend', ['only' => 'resendInvitation']);
     }
 
     /**
      * List User Management
      * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
-    */
+     */
     public function index(Request $request)
     {
-        $data = $this->userManagement->getPaginateUserManagement($request->keyword, $request->status, $request->role, $request->show != null ? $request->show : 10);
-        return $this->sendResponse(new PaginationResource($data), 'Data Fetched Successfully');
+        if ($request->show != null && $request->show != 'all') {
+            $data = $this->userManagement->getPaginateUserManagement($request->keyword, $request->status, $request->role, $request->show != null ? $request->show : 10);
+            $res = new PaginationResource($data);
+        } else {
+            $res = $this->userManagement->getUserManagement($request->keyword, $request->status, $request->role);
+        }
+        return $this->sendResponse($res, 'Data Fetched Successfully');
     }
 
     /**
      * Detail User Management
      * @param \App\Models\UserManager $id
      * @return \Illuminate\Http\Response
-    */
-    public function detail($id)
+     */
+    public function show($id)
     {
-         return $this->sendResponse($this->userManagement->detailUserManagement($id), 'Data Fetched Successfully');
+        return $this->sendResponse($this->userManagement->detailUserManagement($id), 'Data Fetched Successfully');
     }
 
     /**
      * Create or Invite User Management
      * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
-    */
+     */
     public function create(Request $request)
     {
         $validator = Validator::make($request->all(), [
             'user_id'              => 'required',
             'role'                 => 'required',
+            'branch_id'            => '',
         ]);
 
         if ($validator->fails()) {
@@ -83,6 +95,13 @@ class UserManagementController extends BaseController
                 return $this->sendBadRequest('This User Has Been Invited Before');
             }
 
+            if ($request->branch_id != null) {
+                $branch = $request->branch_id;
+            } else {
+                $branch = branchManagerSelected('sanctum:manager')->pivot->branch_id;
+                // return $branch;
+            }
+
             $dataUserManager = [
                 'name'              => $user->name,
                 'email'             => $user->email,
@@ -92,21 +111,21 @@ class UserManagementController extends BaseController
                 'password'          => Hash::make($password),
             ];
             $mailKey = generate_email_verification_key();
-            
-            $userManager = $this->userManagement->inviteUser($dataUserManager, $request->role, $mailKey);
+
+            $this->userManagement->inviteUser($dataUserManager, $branch, $request->role, $mailKey);
             $dataEmail = [
-                'name'      => $userManager->name,
-                'email'     => $userManager->email,
+                'name'      => $user->name,
+                'email'     => $user->email,
                 'password'  => $password,
-                'role'      => $userManager->roles[0]->name,
+                'role'      => $this->roles->detailRoleManager($request->role)->name,
                 'key'       => $mailKey,
             ];
-            Mail::to($userManager->email)->send(new UserManagerVerification($dataEmail));
+            dispatch(new SendEmailJob($dataEmail));
             DB::commit();
-            return $this->sendResponse($userManager, 'User Invited Successfully');
+            return $this->sendResponse($user, 'User Invited Successfully');
         } catch (\Exception $err) {
             DB::rollBack();
-            return $this->sendError(array('success' => false), 'Internal Server Error');
+            return $this->sendError(array('success' => false), $err->getMessage().'-'.$err->getLine());
         }
     }
 
@@ -115,36 +134,47 @@ class UserManagementController extends BaseController
      * @param \Illuminate\Http\Request $request
      * @param \App\Models\UserManager $id
      * @return \Illuminate\Http\Response
-    */
+     */
     public function update(Request $request, $id)
     {
-        $validator = Validator::make($request->all(),[
-            'role'  => 'required',
+        $validator = Validator::make($request->all(), [
+            'role'      => '',
+            'branch_id' => '',
         ]);
         if ($validator->fails()) {
             return $this->sendBadRequest('Validator Error', $validator->errors());
         }
-        $userManager = $this->userManagement->detailUserManagement($id);
-        $userManager->roles()->detach();
-        $userManager->assignRole($request->role);
-        return $this->sendResponse($userManager, 'User Role Updated Successfully');
+        $userManager = $this->userManagement->detailUserBranchAssign($id);
+
+        if ($request->role != null) {
+            $userManager->roles()->detach();
+            $userManager->assignRole($request->role);
+        }
+
+        if ($request->branch_id != null) {
+            $userManager->update([
+                'branch_id' => $request->branch_id,
+            ]);
+        }
+
+        return $this->sendResponse(array('success' => 1), 'User Updated Successfully');
     }
 
     /**
      * Remove or Cancel Invite User Management
      * @param \App\Models\UserManager $userId
      * @return \Illuminate\Http\Response
-    */
-    public function delete($userId)
+     */
+    public function destroy($userId)
     {
-        return $this->sendResponse($this->userManagement->deleteUserManagement($userId), 'User Remove Succesfully');
+        return $this->sendResponse(array('success' => $this->userManagement->deleteUserManagement($userId)), 'User Remove Succesfully');
     }
 
     /**
      * Resend Invite User Management
      * @param \App\Models\UserManager $userId
      * @return \Illuminate\Http\Response
-    */
+     */
     public function resendInvitation($userId)
     {
         $userManager = $this->userManagement->detailUserManagement($userId);
@@ -162,6 +192,4 @@ class UserManagementController extends BaseController
         Mail::to($userManager->email)->send(new UserManagerVerification($dataEmail));
         return $this->sendResponse($userManager, 'User Invited Successfully');
     }
-
-    
 }
